backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
supabase>=2.3.0
anthropic>=0.18.0
openai>=1.12.0
newspaper3k>=0.2.8
lxml>=5.1.0
lxml_html_clean>=0.1.0
python-dotenv>=1.0.0
pydantic>=2.6.0
resend>=0.7.0
apscheduler>=3.10.0
pytz>=2024.1
numpy>=1.26.0
